Correct subsetting rules
Corrects case where user did not want growth form subsetting but
included a list of growth forms, some of which didn’t match (which is
irrelevant) - so only tests matching GFs when requested to subset GFs.

Next, rather than subsetting hits to at or above minimum height
requested, does this for >0 height, and if height is set to 0, it only
excludes hits with no growth form recorded - so that substrate is
excluded but the few cases where there is a plant growth form and no
height recorded, those are included (we don’t know the height but user
does not care in this case).