IVI upgrade
Change calculation of IVI from mean between cover and freq percent to the sum of them each relative to the total across all species (i.e. rowSums of the respective table), so it is the sum of the proportion of cumulative species cover and the proportion of cumulative frequencies and can range from 0 - 200. Also updated the Rd to describe IVI better and add context about the species table use downstream