Update species_optimisation.R
To avoid calling sort() on a 1-row df (rather than numeric vector), when sorting dissimilarities compared to the start plot in simpson_beta_opt. Instead, use sort.list on unnamed vector and then use those positions to re-arranged the 1-row df to get the correct order plus names, which are needed to save as the result (selected plot) See lines 146 ~ 156.