Update growthform_pie.R
When filtering out growth forms with 0 cover before plotting pie chart, change from removing those that are 0 to including all that are > 0. This is because if all growth forms in the growth form table are present, '-which' does not run.

This solves the problem of some ausplots crashing when attempting to run ausplots_visual on just that plot.